tests/wpt/web-platform-tests/referrer-policy/css-integration/child-css/external-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - Child css from external stylesheet</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that child css are loaded with the referrer and referrer policy
    from the external stylesheet.</p>

    <div class="styled"></div>

    <script>
      let css_test = async_test("Child css from external stylesheet.");
      let id = token();
      let url_prefix = location.protocol + "//www1." + location.hostname + ":" + location.port;
      let css_url = url_prefix +
        "/referrer-policy/generic/subresource/stylesheet.py?id=" + id +
        "&import-rule" + "&referrer-policy=no-referrer";
      let check_url = url_prefix + "/referrer-policy/generic/subresource/stylesheet.py" +
                      "?id=" + id + "&report-headers";

      let link = document.createElement("link");
      link.href = css_url;
      link.rel = "stylesheet";
      link.onload = function() {
        css_test.step_timeout(
            queryXhr.bind(this, check_url,
                          function(message) {
                            css_test.step(function() {
                              assert_own_property(message, "headers");
                              assert_equals(message.referrer, undefined);
                            });
                            css_test.done();
                          }),
            1000);
      };
      document.head.appendChild(link);
    </script>

    <div id="log"></div>
  </body>
</html>
